# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clash Ren FF Tournament Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SendOtpBody {
  /** 10-digit phone number (without country code) */
  phone: string;
}

export interface VerifyOtpBody {
  phone: string;
  /** 6-digit OTP code */
  otp: string;
}

export interface User {
  id: number;
  phone?: string | null;
  googleId?: string | null;
  email?: string | null;
  displayName?: string | null;
  avatarUrl?: string | null;
  isProfileComplete: boolean;
  inGameName?: string | null;
  uid?: string | null;
  profilePicture?: string | null;
  diamondBalance: number;
  isAdmin: boolean;
  createdAt: string;
  status?: string;
  blockedAt?: string | null;
  blockedReason?: string | null;
  blockedUntil?: string | null;
  deletedAt?: string | null;
  deleteReason?: string | null;
  allowDepositWithdrawal?: boolean;
  nameChangedAt?: string | null;
  nameChangeAllowed?: boolean;
  twoFaResetAt?: string | null;
  twoFaWithdrawalBypass?: boolean;
  platformId?: string | null;
}

export interface AuthResponse {
  token: string;
  user: User;
  isNewUser: boolean;
}

export interface UpdateUserBody {
  inGameName?: string;
  uid?: string;
  profilePicture?: string | null;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResponseMetadata = {
  name?: string;
  size?: number;
  contentType?: string;
};

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: RequestUploadUrlResponseMetadata;
}

export interface UserStats {
  tournamentsPlayed: number;
  totalKills: number;
  totalWins: number;
  diamondsEarned: number;
  diamondsSpent: number;
  rank?: number | null;
}

export type TournamentStatus =
  (typeof TournamentStatus)[keyof typeof TournamentStatus];

export const TournamentStatus = {
  upcoming: "upcoming",
  ongoing: "ongoing",
  completed: "completed",
} as const;

export interface Tournament {
  id: number;
  title: string;
  gameMode: string;
  entryFeeDiamonds: number;
  prizePoolDiamonds: number;
  maxSlots: number;
  filledSlots: number;
  startTime: string;
  status: TournamentStatus;
  roomId?: string | null;
  roomPassword?: string | null;
  isJoined: boolean;
  createdAt: string;
}

export type CreateTournamentBodyStatus =
  (typeof CreateTournamentBodyStatus)[keyof typeof CreateTournamentBodyStatus];

export const CreateTournamentBodyStatus = {
  upcoming: "upcoming",
  ongoing: "ongoing",
  completed: "completed",
} as const;

export interface CreateTournamentBody {
  title: string;
  gameMode: string;
  entryFeeDiamonds: number;
  prizePoolDiamonds: number;
  maxSlots: number;
  startTime: string;
  status: CreateTournamentBodyStatus;
  roomId?: string;
  roomPassword?: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  inGameName: string;
  profilePicture?: string | null;
  totalKills: number;
  totalWins: number;
  tournamentsPlayed: number;
  diamondsEarned: number;
  /** Points scored (kills*1 + wins*10 + tournamentsPlayed*2) */
  points: number;
}

export interface HistoryEntry {
  id: number;
  tournamentId: number;
  tournamentTitle: string;
  gameMode: string;
  kills: number;
  placement?: number | null;
  diamondsWon: number;
  entryFeeDiamonds: number;
  joinedAt: string;
  tournamentStartTime: string;
  status: string;
}

export interface AdjustDiamondsBody {
  /** Amount to add (positive) or subtract (negative) */
  amount: number;
  reason?: string;
}

export interface AdminStats {
  totalUsers: number;
  totalTournaments: number;
  activeTournaments: number;
  upcomingTournaments: number;
  completedTournaments: number;
  totalDiamondsInCirculation: number;
  totalEntryFeesCollected: number;
  totalPrizesDistributed: number;
}

export interface TournamentParticipantDetail {
  id: number;
  tournamentId: number;
  userId: number;
  inGameName?: string | null;
  phone: string;
  kills: number;
  placement?: number | null;
  diamondsWon: number;
  joinedAt: string;
}

export interface UpdateParticipantBody {
  kills?: number;
  placement?: number | null;
  diamondsWon?: number;
}

export type ListTournamentsParams = {
  status?: ListTournamentsStatus;
};

export type ListTournamentsStatus =
  (typeof ListTournamentsStatus)[keyof typeof ListTournamentsStatus];

export const ListTournamentsStatus = {
  upcoming: "upcoming",
  ongoing: "ongoing",
  completed: "completed",
} as const;

export type GetLeaderboardParams = {
  limit?: number;
  /** Time window: 'weekly' (7 days, default) or 'monthly' (30 days) */
  period?: "weekly" | "monthly";
  /** Game mode filter: 'clash_squad' | 'battle_royal' | 'knockout' */
  mode?: "clash_squad" | "battle_royal" | "knockout";
};