# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clash Ren FF Tournament Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpBody"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: OTP verified, returns auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-otp:
    post:
      operationId: resendOtp
      tags: [auth]
      summary: Resend OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpBody"
      responses:
        "200":
          description: OTP resent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SendOtpBody:
      type: object
      properties:
        phone:
          type: string
          description: 10-digit phone number (without country code)
      required:
        - phone
    VerifyOtpBody:
      type: object
      properties:
        phone:
          type: string
        otp:
          type: string
          description: 6-digit OTP code
      required:
        - phone
        - otp
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        isNewUser:
          type: boolean
      required:
        - token
        - user
        - isNewUser
    User:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
          nullable: true
        googleId:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isProfileComplete:
          type: boolean
        inGameName:
          type: string
          nullable: true
        uid:
          type: string
          nullable: true
        profilePicture:
          type: string
          nullable: true
        diamondBalance:
          type: integer
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - isProfileComplete
        - diamondBalance
        - isAdmin
        - createdAt
...